+++ kernel/src/main.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright 2022 The unCORE Kernel Organization
+
 // ? GLOBAL CRATE ATTRIBUTES AND DOCUMENTATION
 // ? ---------------------------------------------------------------------
 
@@ -31,7 +34,7 @@
 #![feature(custom_test_frameworks)]
 // With our own test framework, we have to define which function
 // runs our tests.
-#![test_runner(library::test_runner)]
+#![test_runner(kernel::prelude::test::runner)]
 // We will have to re-export the actual test runner above with
 // a new name so cargo is not confused.
 #![reexport_test_harness_main = "__test_runner"]
@@ -47,41 +50,36 @@
 // ? MODULES and GLOBAL / CRATE-LEVEL FUNCTIONS
 // ? ---------------------------------------------------------------------
 
-/// # Imports
-///
-/// The `kernel::library` is used here explicitly with the `use`
-/// statement, and not with the `mod` statement. As `kernel::library`
-/// is already used in `lib.rs`, we do not want to re-import it here
-/// and possibly confuse Cargo.
-///
-/// The only exceptions so far is the `init()` function called at the
-/// beginning of `_start`. It is called vi a`kernel::init()` which is
-/// perfectly fine.
-///
-/// ## Macros
-///
-/// We will need to re-import all needed macros, as per definition
-/// they reside in `crate`, which to be exact is `lib.rs`'s root and
-/// not `main.rs`'s root.
-///
-/// Make sure to **always** use `library::` instead of `crate::lib::`
-/// or `lib::` or something else.
-use kernel::library;
+use kernel::{
+	library,
+	prelude::*,
+};
 
-/// # Kernel Binary Entrypoint
+/// ### Kernel Binary Entrypoint
 ///
-/// This is the kernel's entry point called after the bootloader has
-/// finished its setup. It is kept short on purpose. The
-/// `library::init()` function takes care of initialization.
+/// This is the kernel's entry point directly called by the boot-code
+/// (written in assembly). We're still in the UEFI boot services are
+/// still enabled: it is our job to disable them now.
 #[no_mangle]
-pub extern "C" fn _start(boot_information: &'static mut bootloader::BootInfo) -> !
+pub fn kernel_main(
+	multiboot2_bootloader_magic_value: u32,
+	multiboot2_boot_information_pointer: u32,
+) -> !
 {
-	library::init(boot_information);
-
 	#[cfg(test)]
 	__test_runner();
 
-	library::never_return()
+	library::log::init(Some(log::Level::Trace));
+	library::log::display_initial_information();
+
+	library::boot::check_and_parse_multiboot2(
+		multiboot2_bootloader_magic_value,
+		multiboot2_boot_information_pointer,
+	);
+	// https://github.com/rust-osdev/bootloader/blob/main/src/bin/uefi.rs#L37
+	let _uefi_memory_map = library::boot::exit_uefi_boot_services();
+
+	never_return()
 }
 
 /// ### Default Panic Handler
@@ -90,4 +88,4 @@ pub extern "C" fn _start(boot_information: &'static mut bootloader::BootInfo) ->
 /// on whether you are running tests or not, writes an exit code and
 /// does not return afterwards. Note that we do not unwind the stack.
 #[panic_handler]
-fn panic(panic_info: &::core::panic::PanicInfo) -> ! { library::panic_callback(false, panic_info) }
+fn panic(panic_info: &::core::panic::PanicInfo) -> ! { panic_callback(false, panic_info) }