+++ kernel/tests/basic_boot.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright 2022 The unCORE Kernel Organization
+
 // ? GLOBAL CRATE ATTRIBUTES AND DOCUMENTATION
 // ? ---------------------------------------------------------------------
 
@@ -20,36 +23,31 @@
 #![feature(custom_test_frameworks)]
 // With our own test framework, we have to define which function
 // runs our tests.
-#![test_runner(library::test_runner)]
+#![test_runner(test::runner)]
 // We will have to re-export the actual test runner above with
 // a new name so cargo is not confused.
 #![reexport_test_harness_main = "__test_runner"]
 
-use kernel::library::{
-	self,
-	helper::log,
-};
-
 // ? MODULES and GLOBAL / CRATE-LEVEL FUNCTIONS
 // ? ---------------------------------------------------------------------
 
+use kernel::prelude::*;
+
 #[no_mangle]
-pub extern "C" fn _start(boot_information: &'static mut bootloader::BootInfo) -> !
+pub extern "C" fn _start(__todo: u32) -> !
 {
-	log::set_log_level(log::Level::Trace);
-	kernel::log!("Running an integration test.");
-	library::init(boot_information);
+	// test::main(None, boot_information);
 
 	__test_runner();
 
-	library::never_return()
+	never_return()
 }
 
 #[panic_handler]
-fn panic(panic_info: &::core::panic::PanicInfo) -> ! { library::panic_callback(false, panic_info) }
+fn panic(panic_info: &::core::panic::PanicInfo) -> ! { panic_callback(false, panic_info) }
 
 #[test_case]
 fn test_println()
 {
-	kernel::log_info!("Test log output. Does not panic.");
+	log_info!("Test log output. Does not panic.");
 }